Add new branches in the celltype hierarchy added for dynamic nodata

+++ raster/src/main/scala/geotrellis/raster/io/geotiff/ByteGeoTiffSegment.scala
@@ -9,10 +9,10 @@ import java.util.BitSet
 import spire.syntax.cfor._
 
 
-class ByteGeoTiffSegment(bytes: Array[Byte], noDataValue: Byte) extends RawByteGeoTiffSegment(bytes) {
+class ByteGeoTiffSegment(bytes: Array[Byte], noDataValue: Double) extends RawByteGeoTiffSegment(bytes) {
   override
   def get(i: Int): Byte =
-    if (bytes(i) == noDataValue)
+    if (bytes(i) == noDataValue.toByte)
       byteNODATA
     else
       bytes(i)
